+++ next.config.js
@@ -1,3 +1,11 @@
 module.exports = {
   reactStrictMode: true,
+  async redirects() {
+    return [
+      {
+        source: '/_error',
+        destination: '/404'
+      }
+    ]
+  }
 }
